remove bridge macro
